requirements.txt
aiosmtplib==2.0.2
annotated-types==0.5.0
anyio==3.7.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
cryptography==41.0.2
dnspython==2.4.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.0
fastapi-mail==1.4.1
graphql-core==3.2.3
h11==0.14.0
httpcore==0.17.3
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
motor
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==2.0.3
pydantic-settings==2.0.2
pydantic_core==2.3.0
pymongo==3.11.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
strawberry-graphql==0.195.2
typing_extensions==4.7.1
